Update view to their respective views.py
